Converted All SPH Field Methods to float precision methods

+++ kernel/cubic/cubic.go
@@ -1,24 +1,24 @@
 package kernel
 
-import V "dslfluid.com/dsl/math/math64"
+import V "dslfluid.com/dsl/math/math32"
 
 const PI = 3.141592653589
 
 type Cubic struct {
-	A  float64
-	H  float64
-	H0 float64
+	A  float32
+	H  float32
+	H0 float32
 }
 
 //Build Cubic builds cubic kernel function
-func Build_Cubic(h float64) Cubic {
+func Build_Cubic(h float32) Cubic {
 	cubic := Cubic{0, h, h}
 	cubic.A = 1 / (PI * h * h * h)
 	return cubic
 }
 
 //**-------------------CUBIC KERNEL BSPLINE----------------------//
-func (K Cubic) F(x float64) float64 {
+func (K Cubic) F(x float32) float32 {
 	r := x / K.H0
 	if r > 2.0 {
 		return 0.0
@@ -34,12 +34,12 @@ func (K Cubic) F(x float64) float64 {
 	return K.A * 0.25 * s * s * s
 }
 
-func (K Cubic) W0() float64 {
+func (K Cubic) W0() float32 {
 	return K.F(0)
 }
 
 //O1D - 1st order Differential
-func (K Cubic) O1D(x float64) float64 {
+func (K Cubic) O1D(x float32) float32 {
 	//Try the functional derivative
 	r := x / K.H0
 	q := (2 - r)
@@ -56,7 +56,7 @@ func (K Cubic) O1D(x float64) float64 {
 }
 
 //O2D - Returns 2nd Order Differential
-func (K Cubic) O2D(x float64) float64 {
+func (K Cubic) O2D(x float32) float32 {
 
 	//Try the functional derivative
 	r := x / K.H0
@@ -76,7 +76,7 @@ func (K Cubic) O2D(x float64) float64 {
 }
 
 //Adjust changes the kernel smoothing length based on density ratio
-func (K Cubic) Adjust(densityRatio float64) float64 {
+func (K Cubic) Adjust(densityRatio float32) float32 {
 
 	K.H0 = K.H
 
@@ -84,6 +84,6 @@ func (K Cubic) Adjust(densityRatio float64) float64 {
 }
 
 //Grad finds the kernel gradient
-func (K Cubic) Grad(x float64, dir V.Vec) V.Vec {
+func (K Cubic) Grad(x float32, dir V.Vec) V.Vec {
 	return V.Scl(dir, -K.O1D(x))
 }
